Include ResourceService as provider for projectdata
