+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/resource/project/ProjectResourceService.groovy
@@ -1,6 +1,11 @@
 package life.qbic.portal.sampletracking.resource.project
 
 import life.qbic.datamodel.dtos.projectmanagement.Project
+import life.qbic.datamodel.dtos.projectmanagement.ProjectCode
+import life.qbic.datamodel.dtos.projectmanagement.ProjectIdentifier
+import life.qbic.datamodel.dtos.projectmanagement.ProjectSpace
+import life.qbic.datamodel.samples.Location
+import life.qbic.datamodel.samples.Sample
 import life.qbic.portal.sampletracking.communication.Topic
 import life.qbic.portal.sampletracking.resource.ResourceService
 
@@ -11,9 +16,13 @@ import life.qbic.portal.sampletracking.resource.ResourceService
  */
 class ProjectResourceService extends ResourceService<Project> {
 
+    private final List<Project> projects
+
     ProjectResourceService() {
         addTopic(Topic.PROJECT_ADDED)
         addTopic(Topic.PROJECT_REMOVED)
+        this.projects = new ArrayList<Project>()
+        generateMockData()
     }
 
     /**
@@ -36,5 +45,23 @@ class ProjectResourceService extends ResourceService<Project> {
     @Override
     void removeFromResource(Project resourceItem) {
         publish(resourceItem, Topic.PROJECT_REMOVED)
+    }
+
+    void generateMockData(){
+        Location location = new Location()
+
+        Sample sample = new Sample()
+        sample.setCode("Q1234AC")
+        sample.setCurrentLocation(location)
+
+        Project project1 = new Project.Builder(new ProjectIdentifier(new ProjectSpace("My Awesome ProjectSpace 1"), new ProjectCode("QABCD")), "My Awesome Project1").build()
+        Project project2 = new Project.Builder(new ProjectIdentifier(new ProjectSpace("My Awesome ProjectSpace 2"), new ProjectCode("QABCE")), "My Awesome Project2").build()
+        Project project3 = new Project.Builder(new ProjectIdentifier(new ProjectSpace("My Awesome ProjectSpace 3"), new ProjectCode("QABCF")), "My Awesome Project3").build()
+        projects.addAll(project1,project2,project3)
+    }
+
+    @Override
+    Iterator<Project> iterator() {
+        return new ArrayList(projects).iterator()
     }
 }
